refactor: move xsd pipeline under readers/ and proto under writers/
