+++ src/lib.rs
@@ -1,9 +1,7 @@
 pub mod cli;
 pub mod error;
-pub mod proto;
-pub mod resolver;
-pub mod transform;
-pub mod xsd;
+pub mod readers;
+pub mod writers;
 
 #[cfg(test)]
 pub mod test_fixtures;
