+++ src/readers/xsd/mod.rs
@@ -1,3 +1,5 @@
 pub mod model;
 pub mod names;
 pub mod parser;
+pub mod resolver;
+pub mod transform;
