+++ src/readers/xsd/transform/naming.rs
