+++ src/writers/proto/emitter.rs
